Propogate context to wrapped DoFn in DoFns.detach()
When wrapping a DoFn in DoFns.detach(), the context does not get passed down to the
enclosing DoFn when setContext is invoked, meaning counter incrementing and other
features which rely on the context being available don't work as expected. This patch
makes setContext change the context of both the wrapper and the delegate DoFn.